Update SearchResults component
